fix(blog): api
