+++ src/lib/markdown.server.ts
@@ -17,25 +17,22 @@ type Metadata = {
 	isPublished: boolean;
 };
 
-export function parseMarkdown(
-	filepath: string,
-	contentRaw: string,
-): (Item & { content: string }) {
-	const filename = filepath.split('/').at(-1);
+export async function parseMarkdown(
+	slug: string,
+): Promise<Item & { content: string }> {
+	const posts = import.meta.glob(`$posts/*.md`, { eager: true, as: 'raw' });
 
-	/** if not md file, throw error */
-	if (filename != null && !filename.endsWith('.md')) {
-		throw new Error('File is not a markdown file');
+	const [_, mdRaw] = Object.entries(posts).find(([filepath]) => filepath.endsWith(`/${slug}.md`)) ?? [];
+	if (mdRaw == null) {
+		throw new Error(`Post not found: ${slug}`);
 	}
 
-	const slug = filename?.replace('.md', '');
-
-	const { data: metadata, content } = matter(contentRaw);
+	const { data: metadata, content } = matter(mdRaw);
 
 	typia.assertGuard<Metadata>(metadata);
 
 	const item = {
-		slug: slug as string,
+		slug,
 		title: metadata.title,
 		isPublished: metadata.isPublished,
 		pubDate: parse(metadata.date, 'yyyy-MM-dd', new Date()).toJSON(),
