--- src/routes/blog/+page.ts
@@ -1,7 +1,0 @@
-import { posts as macroPosts } from '$lib/api' with { type: 'macro' };
-
-export function load() {
-	return {
-		posts: macroPosts,
-	};
-}
